refactor: move controller package to gitops package
