+++ internal/gitops/controller/compose_pipe.go
