+++ internal/gitops/controller/pruner.go
