+++ internal/gitops/controller/pruner_test.go
