+++ internal/gitops/controller/rotator.go
