+++ internal/gitops/controller/stack_reconciler.go
